requirements.txt
pytest
pytest-cov
pytest-html
pytest-xdist
requests
allure-pytest
